Update hoofdletter klassen
