--- src/main/java/be/intec/themarujohyperblog/Controller/CommentController.java
@@ -1,4 +1,0 @@
-package be.intec.themarujohyperblog.Controller;
-
-public class CommentController {
-}
