--- src/main/java/be/intec/themarujohyperblog/Service/CommentService.java
@@ -1,4 +1,0 @@
-package be.intec.themarujohyperblog.Service;
-
-public interface CommentService {
-}
