--- src/main/java/be/intec/themarujohyperblog/Service/PostService.java
@@ -1,4 +1,0 @@
-package be.intec.themarujohyperblog.Service;
-
-public interface PostService {
-}
