--- src/main/java/be/intec/themarujohyperblog/Service/UserService.java
@@ -1,4 +1,0 @@
-package be.intec.themarujohyperblog.Service;
-
-public interface UserService {
-}
