+++ src/main/java/be/intec/themarujohyperblog/controller/PostController.java
@@ -1,0 +1,4 @@
+package be.intec.themarujohyperblog.controller;
+
+public class PostController {
+}
