+++ src/main/java/be/intec/themarujohyperblog/model/AuditModel.java
@@ -1,4 +1,4 @@
-package be.intec.themarujohyperblog.Model;
+package be.intec.themarujohyperblog.model;
 
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
