+++ src/main/java/be/intec/themarujohyperblog/service/CommentService.java
@@ -1,0 +1,4 @@
+package be.intec.themarujohyperblog.service;
+
+public interface CommentService {
+}
