+++ src/main/java/be/intec/themarujohyperblog/service/PostService.java
@@ -1,0 +1,4 @@
+package be.intec.themarujohyperblog.service;
+
+public interface PostService {
+}
