+++ src/main/java/be/intec/themarujohyperblog/service/UserService.java
@@ -1,0 +1,4 @@
+package be.intec.themarujohyperblog.service;
+
+public interface UserService {
+}
